pyproject.toml
[project]
name = "evaluation-report-synthesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.6",
    "fastparquet>=2024.11.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pathlib>=1.0.1",
    "pyarrow>=22.0.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "umap-learn>=0.5.11",
]
